Move ShardingSphereSchema to schema package (#7681)
* Update java doc

* Move ShardingSphereSchema to schema package

* Remove ShardingSphereMetaData.name
